move files under etl folder
